refactor: create subfolders to organise app
