+++ src/components/ExpandableText/ExpandableText.css
