+++ src/pages/Landing/Landing.jsx
@@ -2,11 +2,11 @@ import React, { Component } from "react";
 import {
   bool, string, shape, arrayOf, objectOf, func,
 } from "prop-types";
-import Card from "./components/Card/Card";
-import Loader from "./components/Loader/Loader";
-import ChurchPicker from "./ChurchPicker";
-import ApiEndpoint from "./ApiEndpoint";
-import "./style/Landing.css";
+import Card from "../../components/Card/Card";
+import Loader from "../../components/Loader/Loader";
+import ChurchPicker from "../../components/ChurchPicker/ChurchPicker";
+import ApiEndpoint from "../../ApiEndpoint";
+import "./Landing.css";
 
 class Landing extends Component {
   constructor(props) {
