+++ src/pages/Study/components/BiblePopup/BiblePopup.css
