+++ src/pages/Study/components/StudyHeader/StudyHeader.css
