+++ src/pages/Study/components/StudyLeadImage/StudyLeadImage.jsx
